@@ -8,12 +8,12 @@ include_directories(.
         ./test
         $ENV{BOOST_INCLUDE})
 
-set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} -DDEBUG -std=c++14 ") #-DMEM_USAGE_TEST_ -DMEM_USAGE_TEST_
-set (CMAKE_CXX_FLAGS_RELEASE "-O3 -DNDEBUG -march=native -mtune=native -std=c++14 ") #-pg -D__PERF_TEST_ONLY___ -DMEM_USAGE_TEST_
+set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} -DDEBUG -DMEM_USAGE_TEST_ -std=c++17 ") #-DMEM_USAGE_TEST_ -DPERFORMANCE_TEST_
+set (CMAKE_CXX_FLAGS_RELEASE "-O3 -DNDEBUG -march=native -mtune=native -std=c++17 ") #-pg -D__PERF_TEST_ONLY___ -DMEM_USAGE_TEST_
 #set(CMAKE_EXE_LINKER_FLAGS "${CMAKE_EXE_LINKER_FLAGS} -Wl,-wrap,malloc -Wl,-wrap,free -Wl,-wrap,realloc")
 
 set(SOURCE_FILES main.cpp ./src/ContBuilder.cpp ./src/hpUtils.cpp ./src/StaticContBuilder.cpp ./src/MemAllocHook.cpp
-        ./test/SuffixTreeTest.cpp ./test/StaticSuffixTreeTest.cpp ./test/memUsageTest.cpp ./test/performanceTest.cpp ./test/testUtils.cpp)
+        ./test/SuffixTreeTest.cpp ./test/StaticSuffixTreeTest.cpp ./test/memUsageTest.cpp ./test/performanceTest.cpp ./test/testUtils.cpp src/ContAllocator.h test/NodeAllocatorTest.cpp)
 
 add_executable(SuffixTree ${SOURCE_FILES})
 
