@@ -4,6 +4,7 @@
 #include <string>
 #include <unordered_map>
 #include "SuffixTree.h"
+#include "ContAllocator.h"
 
 typedef std::string KeyT;
 typedef std::vector<KeyT> Key2IdxT;
@@ -19,11 +20,11 @@ public:
         first_Suffix,
         second_Suffix,
         leaf_Suffix,
-        total_Suffix,
+        total_Suffix
     };
-    typedef suffix_tree::suffix_tree_impl::IndexT ParsedKeyT[total_Suffix];
+    typedef size_t ParsedKeyT[total_Suffix];
 
-    template<SuffixLevel>
+    template<typename AllocatorT, SuffixLevel>
     struct NodeTraits{
         typedef void KeyTypeT;
         typedef void ParentNodeT;
@@ -40,7 +41,7 @@ public:
             char delimeter = '-');
     ~ContBuilder();
 
-    size_t levels()const;
+    size_t levels()const noexcept;
     bool parseKey(
             const KeyT &key, 
             ContBuilder::ParsedKeyT &res)const;
@@ -62,15 +63,15 @@ protected:
             size_t level, 
             const KeyT &key, 
             size_t startIdx, 
-            size_t endIdx, 
-            suffix_tree::suffix_tree_impl::IndexT &index)const;
+            size_t endIdx,
+            size_t &index)const;
 
     void getNewKeyIndex(
             size_t level, 
             const KeyT &key, 
             size_t startIdx, 
-            size_t endIdx, 
-            suffix_tree::suffix_tree_impl::IndexT &index);
+            size_t endIdx,
+            size_t &index);
 
 private:
     typedef std::vector<Key2IndexT> MetaDataPerLevelsT;
@@ -79,33 +80,89 @@ private:
     char delimeter_;
 };
 
-template<>
-struct ContBuilder::NodeTraits<ContBuilder::root_Suffix>{
+template<typename AllocatorT>
+struct ContBuilder::NodeTraits<AllocatorT, ContBuilder::root_Suffix>{
     typedef std::string KeyTypeT;
     typedef void ParentNodeT;
-    typedef suffix_tree::suffix_tree_impl::SuffixNode<ContBuilder, first_Suffix> ChildNodeT;
-    typedef suffix_tree::suffix_tree_impl::RootNode<ContBuilder> NodeTypeT;
+    typedef suffix_tree::suffix_tree_impl::SuffixNode<ContBuilder, first_Suffix, AllocatorT> ChildNodeT;
+    typedef suffix_tree::suffix_tree_impl::RootNode<ContBuilder, AllocatorT> NodeTypeT;
 };
-template<>
-struct ContBuilder::NodeTraits<ContBuilder::first_Suffix>{
+template<typename AllocatorT>
+struct ContBuilder::NodeTraits<AllocatorT, ContBuilder::first_Suffix>{
     typedef std::string KeyTypeT;
-    typedef suffix_tree::suffix_tree_impl::RootNode<ContBuilder> ParentNodeT;
-    typedef suffix_tree::suffix_tree_impl::SuffixNode<ContBuilder, second_Suffix>  ChildNodeT;
-    typedef suffix_tree::suffix_tree_impl::SuffixNode<ContBuilder, first_Suffix> NodeTypeT;
+    typedef suffix_tree::suffix_tree_impl::RootNode<ContBuilder, AllocatorT> ParentNodeT;
+    typedef suffix_tree::suffix_tree_impl::SuffixNode<ContBuilder, second_Suffix, AllocatorT>  ChildNodeT;
+    typedef suffix_tree::suffix_tree_impl::SuffixNode<ContBuilder, first_Suffix, AllocatorT> NodeTypeT;
 };
-template<>
-struct ContBuilder::NodeTraits<ContBuilder::second_Suffix>{
+template<typename AllocatorT>
+struct ContBuilder::NodeTraits<AllocatorT, ContBuilder::second_Suffix>{
     typedef std::string KeyTypeT;
-    typedef suffix_tree::suffix_tree_impl::SuffixNode<ContBuilder, first_Suffix> ParentNodeT;
-    typedef suffix_tree::suffix_tree_impl::LeafNode<ContBuilder, ValueT> ChildNodeT;
-    typedef suffix_tree::suffix_tree_impl::SuffixNode<ContBuilder, second_Suffix> NodeTypeT;
+    typedef suffix_tree::suffix_tree_impl::SuffixNode<ContBuilder, first_Suffix, AllocatorT> ParentNodeT;
+    typedef suffix_tree::suffix_tree_impl::LeafNode<ContBuilder, ValueT, AllocatorT> ChildNodeT;
+    typedef suffix_tree::suffix_tree_impl::SuffixNode<ContBuilder, second_Suffix, AllocatorT> NodeTypeT;
 };
 
-template<>
-struct ContBuilder::NodeTraits<ContBuilder::leaf_Suffix>{
+template<typename AllocatorT>
+struct ContBuilder::NodeTraits<AllocatorT, ContBuilder::leaf_Suffix>{
     typedef std::string KeyTypeT;
-    typedef suffix_tree::suffix_tree_impl::SuffixNode<ContBuilder, second_Suffix> ParentNodeT;
-    typedef suffix_tree::suffix_tree_impl::LeafNode<ContBuilder, ValueT> NodeTypeT;
+    typedef suffix_tree::suffix_tree_impl::SuffixNode<ContBuilder, second_Suffix, AllocatorT> ParentNodeT;
+    typedef suffix_tree::suffix_tree_impl::LeafNode<ContBuilder, ValueT, AllocatorT> NodeTypeT;
 
     static ValueT defaultValue(){return ValueT();}
 };
+
+template<template<typename T, typename A, typename T::SuffixLevel LvlT> class AllocT>
+class ContNodeAllocators{
+    typedef ContNodeAllocators<AllocT> ThisT;
+public:
+    template<ContBuilder::SuffixLevel lvl>
+    AllocT<ContBuilder, ThisT, lvl> &allocator()noexcept;
+
+    typedef AllocT<ContBuilder, ThisT, ContBuilder::root_Suffix> RootNodeAllocT;
+    typedef AllocT<ContBuilder, ThisT, ContBuilder::first_Suffix> FirstNodeAllocT;
+    typedef AllocT<ContBuilder, ThisT, ContBuilder::second_Suffix> SecondNodeAllocT;
+    typedef AllocT<ContBuilder, ThisT, ContBuilder::leaf_Suffix> LeafNodeAllocT;
+
+    template<ContBuilder::SuffixLevel Level>
+    using AllocatorT = AllocT<ContBuilder, ThisT, Level>;
+
+private:
+    RootNodeAllocT allocRoot_;
+    FirstNodeAllocT allocFirst_;
+    SecondNodeAllocT allocSecond_;
+    LeafNodeAllocT allocLeaf_;
+
+};
+
+using ContNodeAllocatorsT = ContNodeAllocators<suffix_tree::suffix_tree_impl::NodeAllocator>;
+
+template<ContBuilder::SuffixLevel Level>
+using ContAllocatorsT = suffix_tree::suffix_tree_impl::NodeAllocator<ContBuilder, ContNodeAllocatorsT, Level>;
+
+template<>template<>
+inline ContAllocatorsT<ContBuilder::root_Suffix> &
+ContNodeAllocators<suffix_tree::suffix_tree_impl::NodeAllocator>::allocator<ContBuilder::root_Suffix>()noexcept
+{
+    return allocRoot_;
+}
+
+template<>template<>
+inline ContAllocatorsT<ContBuilder::first_Suffix> &
+ContNodeAllocators<suffix_tree::suffix_tree_impl::NodeAllocator>::allocator<ContBuilder::first_Suffix>()noexcept
+{
+    return allocFirst_;
+}
+
+template<>template<>
+inline ContAllocatorsT<ContBuilder::second_Suffix> &
+ContNodeAllocators<suffix_tree::suffix_tree_impl::NodeAllocator>::allocator<ContBuilder::second_Suffix>()noexcept
+{
+    return allocSecond_;
+}
+
+template<>template<>
+inline ContAllocatorsT<ContBuilder::leaf_Suffix> &
+ContNodeAllocators<suffix_tree::suffix_tree_impl::NodeAllocator>::allocator<ContBuilder::leaf_Suffix>()noexcept
+{
+    return allocLeaf_;
+}
