@@ -1,7 +1,14 @@
+#if !defined(MEM_USAGE_TEST_) && !defined(PERFORMANCE_TEST_)
+
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wdeprecated-declarations"
+#pragma GCC diagnostic ignored "-Wterminate"
+#include <boost/test/unit_test.hpp>
+#pragma GCC diagnostic pop
+
 #include "StaticSuffixTree.h"
 #include "StaticContBuilder.h"
 #include "MemAllocHook.h"
-#include <cassert>
 #include <iostream>
 #include <functional>
 #include <string>
@@ -59,214 +66,205 @@ namespace{
 
 }
 
-namespace st_tst{
+BOOST_AUTO_TEST_SUITE( static_suffix_tree_test )
 
 
-    void vanillaTest()
+    BOOST_AUTO_TEST_CASE (vanillaTest)
     {
         StaticContBuilder builder(prepareLevel1Keys(), prepareLevel2Keys(), prepareLevel3Keys(), prepareLevel4Keys());
         st_suffix_tree::StaticSuffixTree<StaticContBuilder, std::string, int> cont(builder);
-        assert(0 == cont.size());
+        BOOST_REQUIRE(0 == cont.size());
         auto it = cont.insert("aaa-bbb-ccc-ddd", 777);
-        assert(1 == cont.size());
-        assert(cont.end() != it);
-        assert(cont.end() != cont.find("aaa-bbb-ccc-ddd"));
-        assert(cont.end() == cont.find("aaa-bbb-cca-ddd"));
+        BOOST_REQUIRE(1 == cont.size());
+        BOOST_REQUIRE(cont.end() != it);
+        BOOST_REQUIRE(cont.end() != cont.find("aaa-bbb-ccc-ddd"));
+        BOOST_REQUIRE(cont.end() == cont.find("aaa-bbb-cca-ddd"));
     }
 
-    void invalidKeyTest()
+    BOOST_AUTO_TEST_CASE (invalidKeyTest)
     {
         StaticContBuilder builder(prepareLevel1Keys(), prepareLevel2Keys(), prepareLevel3Keys(), prepareLevel4Keys());
         st_suffix_tree::StaticSuffixTree<StaticContBuilder, std::string, int> cont(builder);
-        assert(0 == cont.size());
+        BOOST_REQUIRE(0 == cont.size());
         auto it = cont.insert("aaa-bbb-XXX-dda", 777);
-        assert(0 == cont.size());
-        assert(cont.end() == it);
-        assert(cont.end() == cont.find("aaa-bbb-XXX-dda"));
+        BOOST_REQUIRE(0 == cont.size());
+        BOOST_REQUIRE(cont.end() == it);
+        BOOST_REQUIRE(cont.end() == cont.find("aaa-bbb-XXX-dda"));
     }
 
-    void findNonexistTest()
+    BOOST_AUTO_TEST_CASE (findNonexistTest)
     {
         StaticContBuilder builder(prepareLevel1Keys(), prepareLevel2Keys(), prepareLevel3Keys(), prepareLevel4Keys());
         st_suffix_tree::StaticSuffixTree<StaticContBuilder, std::string, int> cont(builder);
-        assert(0 == cont.size());
+        BOOST_REQUIRE(0 == cont.size());
         auto it = cont.insert("aaa-bbb-ccc-ddd", 777);
-        assert(1 == cont.size());
-        assert(cont.end() != it);
-        assert(cont.end() == cont.find("aaa-bbb-cca-ddd"));
+        BOOST_REQUIRE(1 == cont.size());
+        BOOST_REQUIRE(cont.end() != it);
+        BOOST_REQUIRE(cont.end() == cont.find("aaa-bbb-cca-ddd"));
     }
 
-    void insertDuplicateTest()
+    BOOST_AUTO_TEST_CASE (insertDuplicateTest)
     {
         StaticContBuilder builder(prepareLevel1Keys(), prepareLevel2Keys(), prepareLevel3Keys(), prepareLevel4Keys());
         st_suffix_tree::StaticSuffixTree<StaticContBuilder, std::string, int> cont(builder);
-        assert(0 == cont.size());
+        BOOST_REQUIRE(0 == cont.size());
         auto it = cont.insert("aaa-bbb-ccc-ddd", 777);
-        assert(1 == cont.size());
-        assert(cont.end() != it);
-        assert(777 == it.value());
+        BOOST_REQUIRE(1 == cont.size());
+        BOOST_REQUIRE(cont.end() != it);
+        BOOST_REQUIRE(777 == it.value());
         auto it2 = cont.insert("aaa-bbb-ccc-ddd", 888);
-        assert(1 == cont.size());
-        assert(cont.end() != it2);
-        assert(888 == it2.value());
+        BOOST_REQUIRE(1 == cont.size());
+        BOOST_REQUIRE(cont.end() != it2);
+        BOOST_REQUIRE(888 == it2.value());
     }
 
-    void eraseByKeyTest()
+    BOOST_AUTO_TEST_CASE (eraseByKeyTest)
     {
         StaticContBuilder builder(prepareLevel1Keys(), prepareLevel2Keys(), prepareLevel3Keys(), prepareLevel4Keys());
         st_suffix_tree::StaticSuffixTree<StaticContBuilder, std::string, int> cont(builder);
-        assert(0 == cont.size());
+        BOOST_REQUIRE(0 == cont.size());
         auto it = cont.insert("aaa-bbb-ccc-ddd", 777);
-        assert(1 == cont.size());
-        assert(cont.end() != it);
-        assert(777 == it.value());
+        BOOST_REQUIRE(1 == cont.size());
+        BOOST_REQUIRE(cont.end() != it);
+        BOOST_REQUIRE(777 == it.value());
         cont.erase("aaa-bbb-ccc-ddd");
-        assert(0 == cont.size());
+        BOOST_REQUIRE(0 == cont.size());
     }
 
-    void eraseAgainByKeyTest()
+    BOOST_AUTO_TEST_CASE (eraseAgainByKeyTest)
     {
         StaticContBuilder builder(prepareLevel1Keys(), prepareLevel2Keys(), prepareLevel3Keys(), prepareLevel4Keys());
         st_suffix_tree::StaticSuffixTree<StaticContBuilder, std::string, int> cont(builder);
-        assert(0 == cont.size());
+        BOOST_REQUIRE(0 == cont.size());
         auto it = cont.insert("aaa-bbb-ccc-ddd", 777);
-        assert(1 == cont.size());
-        assert(cont.end() != it);
-        assert(777 == it.value());
+        BOOST_REQUIRE(1 == cont.size());
+        BOOST_REQUIRE(cont.end() != it);
+        BOOST_REQUIRE(777 == it.value());
         it = cont.erase("aaa-bbb-ccc-ddd");
-        assert(0 == cont.size());
-        assert(cont.end() == it);
+        BOOST_REQUIRE(0 == cont.size());
+        BOOST_REQUIRE(cont.end() == it);
         it = cont.erase("aaa-bbb-ccc-ddd");
-        assert(0 == cont.size());
-        assert(cont.end() == it);
+        BOOST_REQUIRE(0 == cont.size());
+        BOOST_REQUIRE(cont.end() == it);
     }
 
-    void eraseByIteratorTest()
+    BOOST_AUTO_TEST_CASE (eraseByIteratorTest)
     {
         StaticContBuilder builder(prepareLevel1Keys(), prepareLevel2Keys(), prepareLevel3Keys(), prepareLevel4Keys());
         st_suffix_tree::StaticSuffixTree<StaticContBuilder, std::string, int> cont(builder);
-        assert(0 == cont.size());
+        BOOST_REQUIRE(0 == cont.size());
         auto it = cont.insert("aaa-bbb-ccc-ddd", 777);
-        assert(1 == cont.size());
-        assert(cont.end() != it);
-        assert(777 == it.value());
+        BOOST_REQUIRE(1 == cont.size());
+        BOOST_REQUIRE(cont.end() != it);
+        BOOST_REQUIRE(777 == it.value());
 
         auto it2 = cont.insert("aaa-bba-cca-dda", 333);
-        assert(2 == cont.size());
-        assert(cont.end() != it2);
-        assert(333 == it2.value());
+        BOOST_REQUIRE(2 == cont.size());
+        BOOST_REQUIRE(cont.end() != it2);
+        BOOST_REQUIRE(333 == it2.value());
 
         auto it3 = cont.insert("aaz-bbz-ccz-ddz", 999);
-        assert(3 == cont.size());
-        assert(cont.end() != it3);
-        assert(999 == it3.value());
+        BOOST_REQUIRE(3 == cont.size());
+        BOOST_REQUIRE(cont.end() != it3);
+        BOOST_REQUIRE(999 == it3.value());
 
         auto itAfterErase = cont.erase(cont.end());
-        assert(3 == cont.size());
-        assert(cont.end() == itAfterErase);
+        BOOST_REQUIRE(3 == cont.size());
+        BOOST_REQUIRE(cont.end() == itAfterErase);
 
         itAfterErase = cont.erase(it);
-        assert(2 == cont.size());
-        assert(cont.end() != itAfterErase);
-        assert(999 == itAfterErase.value());
+        BOOST_REQUIRE(2 == cont.size());
+        BOOST_REQUIRE(cont.end() != itAfterErase);
+        BOOST_REQUIRE(999 == itAfterErase.value());
 
         itAfterErase = cont.erase(it3);
-        assert(1 == cont.size());
-        assert(cont.end() == itAfterErase);
+        BOOST_REQUIRE(1 == cont.size());
+        BOOST_REQUIRE(cont.end() == itAfterErase);
 
         itAfterErase = cont.erase(it2);
-        assert(0 == cont.size());
-        assert(cont.end() == itAfterErase);
+        BOOST_REQUIRE(0 == cont.size());
+        BOOST_REQUIRE(cont.end() == itAfterErase);
     }
 
-    void setByIteratorTest()
+    BOOST_AUTO_TEST_CASE (setByIteratorTest)
     {
         StaticContBuilder builder(prepareLevel1Keys(), prepareLevel2Keys(), prepareLevel3Keys(), prepareLevel4Keys());
         st_suffix_tree::StaticSuffixTree<StaticContBuilder, std::string, int> cont(builder);
-        assert(0 == cont.size());
+        BOOST_REQUIRE(0 == cont.size());
         auto it = cont.insert("aaa-bbb-ccc-ddd", 777);
-        assert(1 == cont.size());
-        assert(cont.end() != it);
-        assert(777 == *it);
+        BOOST_REQUIRE(1 == cont.size());
+        BOOST_REQUIRE(cont.end() != it);
+        BOOST_REQUIRE(777 == *it);
 
         *it = 7778;
-        assert(7778 == *it);
+        BOOST_REQUIRE(7778 == *it);
         auto it1 = cont.find("aaa-bbb-ccc-ddd");
-        assert(cont.end() != it1);
-        assert(7778 == *it1);
+        BOOST_REQUIRE(cont.end() != it1);
+        BOOST_REQUIRE(7778 == *it1);
     }
 
-    void copyEmptyContainerTest()
+    BOOST_AUTO_TEST_CASE (copyEmptyContainerTest)
     {
         StaticContBuilder builder(prepareLevel1Keys(), prepareLevel2Keys(), prepareLevel3Keys(), prepareLevel4Keys());
         st_suffix_tree::StaticSuffixTree<StaticContBuilder, std::string, int> cont(builder);
-        assert(0 == cont.size());
+        BOOST_REQUIRE(0 == cont.size());
 
         st_suffix_tree::StaticSuffixTree<StaticContBuilder, std::string, int> contCopy(cont);
-        assert(0 == contCopy.size());
+        BOOST_REQUIRE(0 == contCopy.size());
 
         auto it = cont.insert("aaa-bbb-ccc-ddd", 777);
-        assert(1 == cont.size());
-        assert(cont.end() != it);
-        assert(777 == *it);
+        BOOST_REQUIRE(1 == cont.size());
+        BOOST_REQUIRE(cont.end() != it);
+        BOOST_REQUIRE(777 == *it);
 
         auto it1 = contCopy.find("aaa-bbb-ccc-ddd");
-        assert(contCopy.end() == it1);
+        BOOST_REQUIRE(contCopy.end() == it1);
     }
 
-    void copyContainerTest()
+    BOOST_AUTO_TEST_CASE (copyContainerTest)
     {
         StaticContBuilder builder(prepareLevel1Keys(), prepareLevel2Keys(), prepareLevel3Keys(), prepareLevel4Keys());
         st_suffix_tree::StaticSuffixTree<StaticContBuilder, std::string, int> cont(builder);
-        assert(0 == cont.size());
+        BOOST_REQUIRE(0 == cont.size());
         auto it2 = cont.insert("aaa-bbb-ccc-ddd", 777);
         auto it1 = cont.insert("aaa-bba-cca-dda", 1111);
         auto it3 = cont.insert("aaz-bba-cca-dda", 8888);
         auto it4 = cont.insert("aaz-bbz-ccz-ddz", 99);
-        assert(4 == cont.size());
-        assert(cont.end() != it1);
-        assert(1111 == *it1);
-        assert(cont.end() != it2);
-        assert(777 == *it2);
-        assert(cont.end() != it3);
-        assert(8888 == *it3);
-        assert(cont.end() != it4);
-        assert(99 == *it4);
+        BOOST_REQUIRE(4 == cont.size());
+        BOOST_REQUIRE(cont.end() != it1);
+        BOOST_REQUIRE(1111 == *it1);
+        BOOST_REQUIRE(cont.end() != it2);
+        BOOST_REQUIRE(777 == *it2);
+        BOOST_REQUIRE(cont.end() != it3);
+        BOOST_REQUIRE(8888 == *it3);
+        BOOST_REQUIRE(cont.end() != it4);
+        BOOST_REQUIRE(99 == *it4);
 
+        /// make a copy
         st_suffix_tree::StaticSuffixTree<StaticContBuilder, std::string, int> contCopy(cont);
-        assert(4 == contCopy.size());
+        BOOST_REQUIRE(4 == contCopy.size());
+
+        /// update values in original container
         *it1 = 11112;
         *it2 = 7772;
         *it3 = 88882;
         *it4 = 992;
 
+        /// copy of original contains old values
         auto cit = contCopy.find("aaa-bba-cca-dda");
-        assert(contCopy.end() != cit);
-        assert(1111 == *cit);
+        BOOST_REQUIRE(contCopy.end() != cit);
+        BOOST_REQUIRE(1111 == *cit);
         cit = contCopy.find("aaa-bbb-ccc-ddd");
-        assert(contCopy.end() != cit);
-        assert(777 == *cit);
+        BOOST_REQUIRE(contCopy.end() != cit);
+        BOOST_REQUIRE(777 == *cit);
         cit = contCopy.find("aaz-bba-cca-dda");
-        assert(contCopy.end() != cit);
-        assert(8888 == *cit);
+        BOOST_REQUIRE(contCopy.end() != cit);
+        BOOST_REQUIRE(8888 == *cit);
         cit = contCopy.find("aaz-bbz-ccz-ddz");
-        assert(contCopy.end() != cit);
-        assert(99 == *cit);
+        BOOST_REQUIRE(contCopy.end() != cit);
+        BOOST_REQUIRE(99 == *cit);
     }
 
+BOOST_AUTO_TEST_SUITE_END()
 
-    void staticContainerTestSuite()
-    {
-        vanillaTest();
-        invalidKeyTest();
-        findNonexistTest();
-        insertDuplicateTest();
-        eraseByKeyTest();
-        eraseAgainByKeyTest();
-        eraseByIteratorTest();
-        setByIteratorTest();
-        copyEmptyContainerTest();
-        copyContainerTest();
-    }
-
-}
+#endif
